requirements.txt
flask
flask-cors
requests
python-dotenv
openai
pypdf
python-docx
boto3
botocore
aws-wsgi
